updates to play implicits (all colocated inside companion object for SearchResponse), slight reorg
