+++ src/main/scala/discogs/client/DiscogsClient.scala
@@ -2,20 +2,15 @@ package discogs.client
 
 import gigahorse._
 import support.apachehttp.Gigahorse
-import play.api.libs.json.JsonConfiguration.Aux
 import play.api.libs.json._
 
-import discogs.fetch.UserTokenFetcher
-import discogs.models._
+import discogs.fetch.{SearchResponse, UserTokenFetcher}
 
 
 trait DiscogsClient {
   def search(query: String): SearchResponse
   def responseParse(responseString: String): SearchResponse = {
-    implicit val cfg: Aux[Json.MacroOptions] = JsonConfiguration(
-      discriminator = "type",
-      typeNaming = JsonNaming(_.toLowerCase)
-    )
+
     val searchResult: JsResult[SearchResponse] =
       Json.parse(responseString).validate[SearchResponse]
 
@@ -23,11 +18,10 @@ trait DiscogsClient {
       case s: JsSuccess[SearchResponse] => s.get
       case e: JsError => SearchResponse(None)
     }
-
   }
 }
 
-class UserTokenDiscogsClient(private val token: String) extends DiscogsClient {
+class UserTokenDiscogsClient(final private[this] val token: String) extends DiscogsClient {
   def search(query: String): SearchResponse = {
     Gigahorse.withHttp(Gigahorse.config) { http =>
       val fetcher = UserTokenFetcher(token, http)
